bench(taos): move perf tests to benches
