+++ taos/Cargo.toml
@@ -51,6 +51,38 @@ name = "benchmark"
 harness = false
 name = "query"
 
+[[bench]]
+harness = false
+name = "stmt"
+
+[[bench]]
+harness = false
+name = "stmt_interlace"
+
+[[bench]]
+harness = false
+name = "stmt2"
+
+[[bench]]
+harness = false
+name = "stmt2_interlace"
+
+[[bench]]
+harness = false
+name = "sql"
+
+[[bench]]
+harness = false
+name = "sql2"
+
+[[bench]]
+harness = false
+name = "sql_ws_interlace"
+
+[[bench]]
+harness = false
+name = "sql_native_interlace"
+
 [features]
 deadpool = ["taos-query/deadpool"]
 default = ["ws-rustls", "optin", "deadpool"]
